Structure of the final project
Goodmorning Mr Breuvart,

So,

I restructrude the hierarchy of the project and i've re-read the documentation to do the work of the "Responsable du modele
memoire", cause it's kinda LLM sourced if you see what i mean ;).

So yeah today i've put the base and i tried to reach out by hitting him up on the phone, but it was in vain. So i had to
take matters into my own hands.